Unrevert trivial things.
