+++ libs/packaging/packaging/src/test/kotlin/net/corda/libs/packaging/internal/v2/CpkLoaderV2Test.kt
@@ -1,5 +1,6 @@
+package net.corda.libs.packaging.internal.v2
+
 import net.corda.libs.packaging.PackagingConstants.CPK_FORMAT_VERSION2_MAINBUNDLE_PLACEHOLDER
-import net.corda.libs.packaging.internal.v2.CpkLoaderV2
 import net.corda.libs.packaging.testutils.TestUtils
 import net.corda.libs.packaging.testutils.cpk.TestCpkV2Builder
 import org.junit.jupiter.api.Assertions.assertAll
